+++ frontend/src/business/hooks/useBlocker.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { Location, useBlocker as reactRouterUserBlocker, useNavigate } from 'react-router-dom';
+import { Location, useBlocker as reactRouterUserBlocker } from 'react-router-dom';
 
 import { useExitPopup } from './usePopup/useExitPopup';
 
@@ -11,8 +11,7 @@ type useBlockerParams = {
 
 export function useBlocker({ when, onConfirm, onCancel }: useBlockerParams) {
   const [blockedGoBack, setBlockedGoBack] = useState<boolean>(true);
-
-  const navigate = useNavigate();
+  const [afterRunCallbacks, setAfterRunCallbacks] = useState<(() => void)[]>([]);
 
   const { openExitPopup } = useExitPopup();
 
@@ -40,7 +39,10 @@ export function useBlocker({ when, onConfirm, onCancel }: useBlockerParams) {
 
   useEffect(() => {
     if (!blockedGoBack) {
-      navigate('/');
+      afterRunCallbacks.forEach(callback => {
+        callback();
+      });
+      setAfterRunCallbacks([]);
     }
   }, [blockedGoBack]);
 
@@ -48,8 +50,9 @@ export function useBlocker({ when, onConfirm, onCancel }: useBlockerParams) {
     setBlockedGoBack(true);
   }, []);
 
-  const unblockGoBack = useCallback(() => {
-    setBlockedGoBack(false);
+  const unblockGoBack = useCallback((afterRunCallback: () => void) => {
+    setBlockedGoBack(true);
+    setAfterRunCallbacks(prev => [...prev, afterRunCallback]);
   }, []);
 
   return { blockedGoBack, blockGoBack, unblockGoBack };
