+++ frontend/src/pages/HumanChatPage/ChattingPage.tsx
@@ -1,10 +1,8 @@
-import { useEffect } from 'react';
-import { useNavigate, useOutletContext, useParams } from 'react-router-dom';
+import { useNavigate, useOutletContext } from 'react-router-dom';
 
 import { IconButton } from '@components/Buttons';
 import CamContainer from '@components/CamContainer';
 
-import { useBlocker } from '@business/hooks/useBlocker';
 import useSpeakerHighlighter from '@business/hooks/useSpeakerHighlighter';
 
 import type { OutletContext } from './HumanChatPage';
@@ -19,13 +17,9 @@ export default function ChattingPage() {
     toggleVideo,
     toggleAudio,
     tarotButtonClick,
-    setChatPageState,
+    chatPageState: { joined },
+    unblockGoBack,
   }: OutletContext = useOutletContext();
-
-  const { unblockGoBack } = useBlocker({
-    when: ({ nextLocation }) => nextLocation.pathname === '/' || nextLocation.pathname === '/chat/human',
-    onConfirm: () => navigate('/'),
-  });
 
   useChattingPageChangeVideoTrackJoined();
   useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack });
@@ -36,12 +30,8 @@ export default function ChattingPage() {
   const navigate = useNavigate();
   const goSettingPage = () => navigate('setting');
 
-  useEffect(() => {
-    setChatPageState(prev => ({ ...prev, joined: true }));
-  }, []);
-
   return (
-    <>
+    <div className={`${joined ? '' : 'hidden'}`}>
       <CamContainer
         localVideoRef={localVideoRef}
         remoteVideoRef={remoteVideoRef}
@@ -58,6 +48,6 @@ export default function ChattingPage() {
           onClick={goSettingPage}
         />
       </div>
-    </>
+    </div>
   );
 }
