+++ frontend/src/pages/HumanChatPage/useChattingPageCreateJoinRoomPopup.ts
@@ -6,7 +6,7 @@ import { HumanSocketManager } from '@business/services/SocketManager';
 import { OutletContext } from './HumanChatPage';
 
 interface useChattingPageCreateJoinRoomParams {
-  unblockGoBack: () => void;
+  unblockGoBack: (cb: () => void) => void;
 }
 export function useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack }: useChattingPageCreateJoinRoomParams) {
   const {
@@ -26,20 +26,23 @@ export function useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack }: us
     joinRoom({
       roomName: roomName as string,
       onSuccess: ({ close }) => {
+        navigate('setting');
         close();
       },
       onFull: () => {
-        unblockGoBack();
-        navigate('/');
-        alert('방이 꽉 찼습니다, 첫페이지로 이동합니다.');
+        unblockGoBack(() => {
+          alert('방이 꽉 찼습니다, 첫페이지로 이동합니다.');
+          navigate('/');
+        });
       },
       onFail: () => {
         alert('잘못된 링크거나 비밀번호가 틀렸습니다.');
       },
       onHostExit: () => {
-        unblockGoBack();
-        navigate('/');
-        alert('호스트가 방을 나갔습니다, 첫페이지로 이동합니다.');
+        unblockGoBack(() => {
+          alert('호스트가 방을 나갔습니다, 첫페이지로 이동합니다.11');
+          navigate('/');
+        });
       },
     });
   };
@@ -48,6 +51,7 @@ export function useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack }: us
     createRoom({
       roomName: roomName as string,
       onSuccess: ({ close }) => {
+        navigate('setting');
         close();
       },
       onClose: ({ close }) => {
@@ -58,9 +62,10 @@ export function useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack }: us
   };
 
   const goRootPageWithMessage = () => {
-    unblockGoBack();
-    alert('방이 존재하지 않습니다, 첫페이지로 이동합니다.');
-    navigate('/');
+    unblockGoBack(() => {
+      alert('방이 존재하지 않습니다, 첫페이지로 이동합니다.');
+      navigate('/');
+    });
   };
 
   useEffect(() => {
